Updated version and LAM requiresReload
requiresReload = true will show a button below the LAM panel of the addon, if you changed that setting. You can click it to reload the UI now, and if you try to leave the LAM panel it will ask you if you want to reload the UI now to make the settings apply properly. 
Or else it will revert the settings change!